requirements.txt
paho-mqtt
pandas
scikit-learn
pytest
filelock
python-dotenv
